+++ cohttp-async/src/server.ml
@@ -149,3 +149,14 @@ let create ?max_connections ?backlog ?buffer_age_limit ?(mode = `TCP)
   in
   create_raw ?max_connections ?backlog ?buffer_age_limit ~on_handler_error ~mode
     where_to_listen handle_request
+
+module Expert = struct
+  let create handle_request addr reader writer =
+    let handle_request ~body addr request =
+      handle_request ~body addr request >>| fun r -> `Response r
+    in
+    handle_client handle_request addr reader writer
+
+  let create_with_response_action handle_request addr reader writer =
+    handle_client handle_request addr reader writer
+end